Add feature to disable/enable hanning window
The hanning window function is useful when there is  a chance of leakage
in the FFT results for some scenarios with periodic signals. However,
windows aren't always the best approach and not using them can result in
a better distribution of the FFT results. Therefore the functions
enableFFTWindow() and disableFFTWindow were added. By default, hanning
window is enabled so the library keeps full compatibility and can be
updated without any change.