Added ZeroFFTMagnitude function.
Added new API ZeroFFTMagnitude() that returns a real-valued result that is the length of the complex vector (the square of this is related to power).  The phase is discarded.
Added gapsqrt32 to compute a fast integer square root (avoids floating point, which more than doubles the run-time of the function).
Added new test to compute RMS error on the new function.  Passes.